fix(api_rust): resolver 37 errores clippy (warnings-as-errors)
- dead_code: eliminar `hex_decode_bytes` en external.rs; marcar
  `UserMentionResult` y `EntitySearchResponse` (schemas utoipa)
  con `#[allow(dead_code)]` — son referenciadas sólo por OpenAPI.
- doc_lazy_continuation / doc_overindented_list_items: re-indentar
  comentarios rustdoc en export.rs e issue_pagination.rs.
- needless_borrows_for_generic_args: quitar `&` redundantes en
  `r.assignees.join()`/`r.labels.join()` (export.rs) y en la
  cadena HMAC-SHA256 de SigV4 (s3_presigned_post.rs, 7 sitios).
- unnecessary_cast: eliminar `size_limit as i64` redundante en
  assets.rs (el valor ya es i64).
- needless_update: quitar `..Default::default()` en ActiveModels
  donde todos los campos NOT NULL ya están especificados —
  cycles, estimates(x3), intake, issues(x2), labels, modules,
  webhooks.
- items_after_test_module: mover `mod tests` al final del archivo
  en cleanup.rs y s3.rs.
- invisible_characters: corregir mojibake
  "VersiÃ³n especÃ­fica" → "Versión específica" en pages.rs.
- unnecessary_lazy_evaluations: `unwrap_or_else(|| Struct{..})`
  → `unwrap_or(Struct{..})` (struct literal es barato) en pages.rs.
- large_enum_variant: `#[allow]` en FilteredQuery con justificación
  (hot path = Active, boxear añadiría alloc sin beneficio).
- if_same_then_else: colapsar ramas duplicadas en
  user_profile_issues.rs (`A||B→X, else Y`).
- too_many_arguments: `#[allow]` en `generate_presigned_post` con
  justificación (firma SigV4, agrupar en struct = ruido).
- non_snake_case: renombrar `updated_at__gt` → `updated_at_gt` +
  `#[serde(rename = "updated_at__gt")]` en V2IssuesQuery
  (preserva contrato query-param con clientes Django).

Nota de seguridad: los `..Default::default()` retirados en
ActiveModels estaban todos donde clippy confirmó que todos los
campos ya se setean explícitamente, así que no cambia la
semántica ni deja campos en NotSet.

+++ apps/api_rust/src/routes/assets.rs
@@ -1391,7 +1391,7 @@ pub async fn initiate_project_asset_upload(
         &state.config.aws_secret_access_key,
         &asset_key,
         content_type,
-        size_limit as i64,
+        size_limit,
         UPLOAD_URL_TTL_SECS as i64,
     )
     .map_err(|e| AppError::Internal(anyhow::anyhow!("presigned post error: {e}")))?;
